checkpoint save. builds and runs.
