+++ src/progly/run-transforms.cc
@@ -18,11 +18,19 @@
 
 namespace po = boost::program_options ;
 
+struct spj_query_op {
+  std::string oid ;
+  std::string pool ;
+  std::vector< std::string > select_atts ;
+  std::vector< std::string > from_rels ;
+  std::vector< std::string > where_preds ;
+} ;
 
 void set_rows( std::string ) ;
 void set_col0( std::string ) ;
 void set_col1( std::string ) ;
 void set_col2( std::string ) ;
+void execute_query( spj_query_op ) ;
 
 int main( int argc, char **argv ) {
   std::cout << "in run-transforms..." << std::endl ;
@@ -97,7 +105,25 @@ int main( int argc, char **argv ) {
 
   // execute transform
   std::cout << "row to col transpose=================================" << std::endl ;
-  execute_transform( to ) ;
+  librados::bufferlist inbl, outbl ;
+  ::encode( to, inbl ) ;
+
+  // connect to rados and open pool
+  librados::Rados cluster;
+  cluster.init(NULL);
+  cluster.conf_read_file(NULL);
+  int ret = cluster.connect();
+  checkret(ret, 0);
+  librados::IoCtx ioctx ;
+  ret = cluster.ioctx_create( "tpchflatbuf" , ioctx ) ;
+  checkret( ret, 0 ) ;
+  ret = ioctx.exec( "atable", "tabular", "sky_transform", inbl, outbl ) ;
+  bufferlist::iterator p = outbl.begin();
+  size_t i = p.get_remaining() ;
+  std::cout << i << std::endl ;
+  ret = ioctx.write( "atable_transposed", outbl, i, 0 ) ;
+  ioctx.close() ;
+
   std::cout << "=================================" << std::endl ;
 
   // query the transpose
@@ -110,21 +136,37 @@ int main( int argc, char **argv ) {
   execute_query( print_cols ) ;
   std::cout << "=================================" << std::endl ;
 // ...................................... //
-
-// ...................................... //
   // define transform operation
-  transform_op to_recompose ;
-  to_recompose.oid            = "atable_transposed" ;
-  to_recompose.pool           = "tpchflatbuf" ;
-  to_recompose.transform_type = 0 ; // 0 --> transpose
-  to_recompose.bloffs.push_back( 0 ) ;
-  to_recompose.bloffs.push_back( 1 ) ;
-  to_recompose.bloffs.push_back( 2 ) ;
-  to_recompose.layout = 1 ;
+  transform_op to1 ;
+  to1.oid            = "atable_transposed" ;
+  to1.pool           = "tpchflatbuf" ;
+  to1.transform_type = 0 ; // 0 --> transpose
+  to1.bloffs.push_back( 0 ) ;
+  to1.bloffs.push_back( 1 ) ;
+  to1.bloffs.push_back( 2 ) ;
 
   // execute transform
   std::cout << "row to col transpose=================================" << std::endl ;
-  execute_transform( to_recompose ) ;
+  librados::bufferlist inbl1, outbl1 ;
+  ::encode( to1, inbl1 ) ;
+
+  // connect to rados and open pool
+  librados::Rados cluster1 ;
+  cluster1.init( NULL ) ;
+  cluster1.conf_read_file( NULL ) ;
+  ret = cluster1.connect() ;
+  checkret(ret, 0);
+  librados::IoCtx ioctx1 ;
+  ret = cluster1.ioctx_create( "tpchflatbuf" , ioctx1 ) ;
+  checkret( ret, 0 ) ;
+  ret = ioctx1.exec( "atable_transposed", "tabular", "sky_transform", inbl1, outbl1 ) ;
+  bufferlist::iterator p1 = outbl1.begin();
+  size_t i1 = p1.get_remaining() ;
+  std::cout << "i1 = " << i1 << std::endl ;
+  ret = ioctx1.write( "atable_transposed_transposed", outbl1, i1, 0 ) ;
+  std::cout << "write ret = " << ret << std::endl ;
+  ioctx1.close() ;
+
   std::cout << "=================================" << std::endl ;
 
   // query the transpose
@@ -133,29 +175,10 @@ int main( int argc, char **argv ) {
   print_rows.pool = "tpchflatbuf" ;
 
   // execute query
-  std::cout << "ROW query=================================" << std::endl ;
+  std::cout << "ROWS query=================================" << std::endl ;
   execute_query( print_rows ) ;
   std::cout << "=================================" << std::endl ;
 // ...................................... //
-
-  // connect to rados
-  librados::Rados cluster;
-  cluster.init(NULL);
-  cluster.conf_read_file(NULL);
-  int ret = cluster.connect();
-  checkret(ret, 0);
-
-  // open pool
-  librados::IoCtx ioctx ;
-  ret = cluster.ioctx_create( "tpchflatbuf" , ioctx ) ;
-  checkret( ret, 0 ) ;
-
-  librados::bufferlist inbl, outbl ;
-  ret = ioctx.exec( oid, "tabular", "sky_transform", inbl, outbl ) ;
-
-  std::cout << "ret = " << ret << std::endl ;
-
-  ioctx.close() ;
 
   return 0 ;
 }
@@ -586,3 +609,166 @@ void set_col0( std::string oid ) {
 
   ioctx.close() ;
 }
+
+// ============================== //
+//          EXECUTE QUERY         //
+// ============================== //
+void execute_query( spj_query_op q_op ) {
+  std::cout << "in execute_query..." << std::endl ;
+  std::cout << "q_op.oid         : " << q_op.oid         << std::endl ;
+  std::cout << "q_op.pool        : " << q_op.pool        << std::endl ;
+  std::cout << "q_op.select_atts : " << q_op.select_atts << std::endl ;
+  std::cout << "q_op.from_rels   : " << q_op.from_rels   << std::endl ;
+  std::cout << "q_op.where_preds : " << q_op.where_preds << std::endl ;
+
+  // connect to rados
+  librados::Rados cluster;
+  cluster.init(NULL);
+  cluster.conf_read_file(NULL);
+  int ret = cluster.connect();
+  checkret(ret, 0);
+
+  // open pool
+  librados::IoCtx ioctx ;
+  ret = cluster.ioctx_create( q_op.pool.c_str(), ioctx ) ;
+  checkret( ret, 0 ) ;
+
+  // read bl_seq
+  librados::bufferlist wrapped_bl_seq ;
+  int num_bytes_read = ioctx.read( q_op.oid.c_str(), wrapped_bl_seq, (size_t)0, (uint64_t)0 ) ;
+  std::cout << "num_bytes_read : " << num_bytes_read << std::endl ; 
+
+  ceph::bufferlist::iterator it_wrapped = wrapped_bl_seq.begin() ;
+
+  // ================================================================================ //
+  // display data
+
+  while( it_wrapped.get_remaining() > 0 ) {
+
+    // grab the Root
+    ceph::bufferlist bl ;
+    ::decode( bl, it_wrapped ) ; // this decrements get_remaining by moving iterator
+    const char* fb = bl.c_str() ;
+
+    auto root      = Tables::GetRoot( fb ) ;
+    auto data_type = root->relationData_type() ;
+
+    std::cout << "data_type : " << data_type << std::endl ;
+
+    // process one Root>Rows flatbuffer
+    if( data_type == Tables::Relation_Rows ) {
+      std::cout << "if data_type == Tables::Relation_Rows" << std::endl ;
+
+      auto schema    = root->schema() ;
+      std::cout << "schema : " << std::endl ;
+      for( unsigned int i = 0; i < schema->Length(); i++ )
+        std::cout << (unsigned)schema->Get(i) << std::endl ;
+ 
+      auto rows = static_cast< const Tables::Rows* >( root->relationData() ) ;
+      auto table_name_read = rows->table_name() ;
+
+      auto schema_read     = rows->schema() ;
+      auto nrows_read      = rows->nrows() ;
+      auto ncols_read      = rows->ncols() ;
+      auto rids_read       = rows->RIDs() ;
+      auto rows_data       = rows->data() ; // [ Record ]
+
+      std::cout << "table_name_read->str() : " << table_name_read->str() << std::endl ;
+      std::cout << "schema_read->Length() : " << schema_read->Length() << std::endl ;
+      std::cout << "nrows_read     : " << nrows_read     << std::endl ;
+      std::cout << "ncols_read     : " << ncols_read     << std::endl ;
+
+      // print schema to stdout
+      std::cout << "RID\t" ;
+      for( unsigned int i = 0; i < ncols_read; i++ ) {
+        std::cout << schema_read->Get(i)->str() << "\t" ;
+      }
+      std::cout << std::endl ;
+
+      //auto int_data = static_cast< const Tables::IntData* >( rows_data->Get(0) ) ;
+      //std::cout << int_data->data()->Get(0) << std::endl ;
+      // print data to stdout
+      for( unsigned int i = 0; i < rows_data->Length(); i++ ) {
+        std::cout << rids_read->Get(i) << ":\t" ;
+
+        auto curr_rec           = rows_data->Get(i) ;
+        auto curr_rec_data      = curr_rec->data() ;
+        auto curr_rec_data_type = curr_rec->data_type() ;
+
+        for( unsigned int j = 0; j < curr_rec_data->Length(); j++ ) {
+          // column of ints
+          if( (unsigned)curr_rec_data_type->Get(j) == Tables::Data_IntData ) {
+            //std::cout << "int" << "\t" ;
+            auto int_col_data = static_cast< const Tables::IntData* >( curr_rec_data->Get(j) ) ;
+            std::cout << int_col_data->data()->Get(0) << "\t" ;
+          }
+          // column of floats
+          else if( (unsigned)curr_rec_data_type->Get(j) == Tables::Data_FloatData ) {
+            //std::cout << "float" << "\t" ;
+            auto float_col_data = static_cast< const Tables::FloatData* >( curr_rec_data->Get(j) ) ;
+            std::cout << float_col_data->data()->Get(0) << "\t" ;
+          }
+          // column of strings
+          else if( (unsigned)curr_rec_data_type->Get(j) == Tables::Data_StringData ) {
+            //std::cout << "str" << "\t" ;
+            auto string_col_data = static_cast< const Tables::StringData* >( curr_rec_data->Get(j) ) ;
+            std::cout << string_col_data->data()->Get(0)->str() << "\t" ;
+          }
+          else {
+            std::cout << "execute_query: unrecognized row_data_type " 
+                      << (unsigned)curr_rec_data_type->Get(j) << std::endl ; 
+          }
+        }
+        std::cout << std::endl ;
+      }
+    } // Relation_Rows 
+
+    // process one Root>Col flatbuffer
+    else if( data_type == Tables::Relation_Col ) {
+      std::cout << "else if data_type == Tables::Relation_Col" << std::endl ;
+
+      auto col = static_cast< const Tables::Col* >( root->relationData() ) ;
+      auto col_name_read  = col->col_name() ;
+      auto col_index_read = col->col_index() ;
+      auto nrows_read     = col->nrows() ;
+      auto rids_read      = col->RIDs() ;
+      auto col_data_type  = col->data_type() ;
+      auto col_data       = col->data() ;
+
+      std::cout << "col_name_read->str() : " << col_name_read->str() << std::endl ;
+      std::cout << "col_index_read       : " << (unsigned)col_index_read << std::endl ;
+      std::cout << "nrows_read           : " << (unsigned)nrows_read     << std::endl ;
+      std::cout << "col_data_type        : " << col_data_type     << std::endl ;
+
+      // print data to stdout
+      for( unsigned int i = 0; i < nrows_read; i++ ) {
+        std::cout << rids_read->Get(i) << ":\t" ;
+        // column of ints
+        if( (unsigned)col_data_type == Tables::Data_IntData ) {
+          auto int_col_data = static_cast< const Tables::IntData* >( col_data ) ;
+          std::cout << int_col_data->data()->Get(i) << "\t" ;
+        }
+        // column of floats
+        else if( (unsigned)col_data_type == Tables::Data_FloatData ) {
+          auto float_col_data = static_cast< const Tables::FloatData* >( col_data ) ;
+          std::cout << float_col_data->data()->Get(i) << "\t" ;
+        }
+        // column of strings
+        else if( (unsigned)col_data_type == Tables::Data_StringData ) {
+          auto string_col_data = static_cast< const Tables::StringData* >( col_data ) ;
+          std::cout << string_col_data->data()->Get(i)->str() << "\t" ;
+        }
+        else {
+          std::cout << "unrecognized data_type " << (unsigned)col_data_type << std::endl ; 
+        }
+        std::cout << std::endl ;
+      }
+    } // Relation_Col
+
+    else {
+      std::cout << "unrecognized data_type '" << data_type << "'" << std::endl ;
+    }
+  } // while
+
+  ioctx.close() ;
+} // EXECUTE QUERY
